requirements.txt
Django==1.7.7
django-admin-bootstrapped==2.3.5
django-bootstrap3==5.2.0
django-daterange-filter==1.1.1
django-nose==1.3
django-phonenumber-field==0.7.2
django-tables2==0.15.0
nose==1.3.4
phonenumbers==7.0.4
Pillow==2.7.0
psycopg2==2.6
six==1.9.0
